refactor: JDBC 및 In-memory 이용 service encapsulation
